tests/wpt/tests/html/canvas/element/text/2d.text.drawing.style.fontKerning.with.uppercase.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by /html/canvas/tools/gentest.py. -->
<meta charset="UTF-8">
<title>Canvas test: 2d.text.drawing.style.fontKerning.with.uppercase</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/html/canvas/resources/canvas-tests.js"></script>
<link rel="stylesheet" href="/html/canvas/resources/canvas-tests.css">
<body class="show_output">

<h1>2d.text.drawing.style.fontKerning.with.uppercase</h1>
<p class="desc">Testing basic functionalities of fontKerning for canvas</p>


<p class="output">Actual output:</p>
<canvas id="c" class="output" width="100" height="50"><p class="fallback">FAIL (fallback content)</p></canvas>

<ul id="d"></ul>
<script>
var t = async_test("Testing basic functionalities of fontKerning for canvas");
_addTest(function(canvas, ctx) {

  _assertSame(ctx.fontKerning, "auto", "ctx.fontKerning", "\"auto\"");
  ctx.fontKerning = "Normal";
  _assertSame(ctx.fontKerning, "auto", "ctx.fontKerning", "\"auto\"");
  ctx.fontKerning = "normal";
  _assertSame(ctx.fontKerning, "normal", "ctx.fontKerning", "\"normal\"");
  ctx.fontKerning = "Auto";
  _assertSame(ctx.fontKerning, "normal", "ctx.fontKerning", "\"normal\"");
  ctx.fontKerning = "auto";
  ctx.fontKerning = "noRmal";
  _assertSame(ctx.fontKerning, "auto", "ctx.fontKerning", "\"auto\"");
  ctx.fontKerning = "auto";
  ctx.fontKerning = "NoRMal";
  _assertSame(ctx.fontKerning, "auto", "ctx.fontKerning", "\"auto\"");
  ctx.fontKerning = "auto";
  ctx.fontKerning = "NORMAL";
  _assertSame(ctx.fontKerning, "auto", "ctx.fontKerning", "\"auto\"");

  ctx.fontKerning = "None";
  _assertSame(ctx.fontKerning, "auto", "ctx.fontKerning", "\"auto\"");
  ctx.fontKerning = "none";
  _assertSame(ctx.fontKerning, "none", "ctx.fontKerning", "\"none\"");
  ctx.fontKerning = "Auto";
  _assertSame(ctx.fontKerning, "none", "ctx.fontKerning", "\"none\"");
  ctx.fontKerning = "auto";
  ctx.fontKerning = "nOne";
  _assertSame(ctx.fontKerning, "auto", "ctx.fontKerning", "\"auto\"");
  ctx.fontKerning = "auto";
  ctx.fontKerning = "nonE";
  _assertSame(ctx.fontKerning, "auto", "ctx.fontKerning", "\"auto\"");
  ctx.fontKerning = "auto";
  ctx.fontKerning = "NONE";
  _assertSame(ctx.fontKerning, "auto", "ctx.fontKerning", "\"auto\"");

});
</script>
